# docs/openapi-public.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/analytics/cpq/dashboard": {
      "get": {
        "operationId": "CpqAnalyticsController_getDashboard",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/trends": {
      "get": {
        "operationId": "CpqAnalyticsController_getTrends",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/quote-pipeline": {
      "get": {
        "operationId": "CpqAnalyticsController_getQuotePipeline",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/top-products": {
      "get": {
        "operationId": "CpqAnalyticsController_getTopProducts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/top-accounts": {
      "get": {
        "operationId": "CpqAnalyticsController_getTopAccounts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/sales-rep-performance": {
      "get": {
        "operationId": "CpqAnalyticsController_getSalesRepPerformance",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/forecast": {
      "get": {
        "operationId": "CpqAnalyticsController_getForecast",
        "parameters": [
          {
            "name": "periods",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/conversion-funnel": {
      "get": {
        "operationId": "CpqAnalyticsController_getConversionFunnel",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/win-loss": {
      "get": {
        "operationId": "CpqAnalyticsController_getWinLossAnalysis",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/snapshots": {
      "get": {
        "operationId": "CpqAnalyticsController_getSnapshots",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      },
      "post": {
        "operationId": "CpqAnalyticsController_createSnapshot",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/quotes/trend": {
      "get": {
        "operationId": "CpqAnalyticsController_getQuotesTrend",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/orders/trend": {
      "get": {
        "operationId": "CpqAnalyticsController_getOrdersTrend",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/products/performance": {
      "get": {
        "operationId": "CpqAnalyticsController_getProductPerformance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
    "/api/analytics/cpq/quotes/age-analysis": {
      "get": {
        "operationId": "CpqAnalyticsController_getQuoteAgeAnalysis",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "CPQ Analytics"
        ]
      }
    },
...
    "/api/competitors/analytics/win-loss": {
      "get": {
        "operationId": "CompetitorsController_getWinLossAnalytics",
        "parameters": [
          {
            "name": "competitorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Competitors"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalesOS API",
    "description": "SalesOS is an AI-powered Sales CRM and Revenue Intelligence platform. This API provides programmatic access to CRM data, sales coaching, CPQ, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://www.salesos.org/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    },
...
  }
}